% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoflow-package.R
\docType{package}
\name{nanoflow-package}
\alias{nanoflow}
\alias{nanoflow-package}
\title{nanoflow: end-to-end Oxford Nanopore long-read WGS workflow}
\description{
An orchestration layer over established command-line tools: basecalling
(Dorado), QC/trimming (NanoPlot, Porechop, Filtlong, MultiQC), alignment
(minimap2), optional de novo assembly (Flye/Canu/Wengan + Racon/Medaka +
QUAST), small-variant calling (Clair3/Medaka), SV calling
(Sniffles2/SVIM/NanoVar + SURVIVOR + Spectre), phasing
(WhatsHap/HapCUT2), annotation (SnpEff/AnnotSV), truth benchmarking
(hap.py/Truvari) and HTML reporting.
}
\details{
Start with \code{\link[=write_config_template]{write_config_template()}}, \code{\link[=check_tools]{check_tools()}} and
\code{\link[=run_pipeline]{run_pipeline()}}; \code{vignette("nanoflow")} walks through the synthetic
fixture end to end.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fiona-meinzer/Nanopore_workflow}
  \item Report bugs at \url{https://github.com/fiona-meinzer/Nanopore_workflow/issues}
}

}
\author{
\strong{Maintainer}: Fiona Katharina Meinzer \email{fiona.ewald@stat.uni-muenchen.de}

Authors:
\itemize{
  \item Fiona Katharina Meinzer \email{fiona.ewald@stat.uni-muenchen.de}
}

}
\keyword{internal}
